refactor(pack): refactor xdsl mtu reboot
