+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/reboot/index.js
@@ -1,0 +1,32 @@
+import angular from 'angular';
+import ngOvhTelecomUniverseComponents from '@ovh-ux/ng-ovh-telecom-universe-components';
+import '@uirouter/angularjs';
+import '@ovh-ux/ng-translate-async-loader';
+import 'angular-translate';
+import 'ovh-api-services';
+
+import controller from './reboot.controller';
+import template from './reboot.html';
+
+const moduleName = 'XdslModemReboot';
+
+angular
+  .module(moduleName, [
+    ngOvhTelecomUniverseComponents,
+    'ngTranslateAsyncLoader',
+    'pascalprecht.translate',
+    'ovh-api-services',
+    'ui.router',
+  ])
+  .controller('XdslModemRebootCtrl', controller)
+  .run(/* @ngTranslationsInject:json ./translations */)
+  .run(
+    /* @ngInject */ ($templateCache) => {
+      $templateCache.put(
+        'app/telecom/pack/xdsl/modem/reboot/pack-xdsl-modem-reboot.html',
+        template,
+      );
+    },
+  );
+
+export default moduleName;
