+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/reboot/reboot.html
